made supply a double due to units taking 0.5 supply. also changed resourcemanager

+++ src/CSVParser.cpp
@@ -38,7 +38,7 @@ void CSVParser::parseEntityLine(string line) {
 	getline(iss, token, ',');
 	entityData.buildTime = stoi(token);
 	getline(iss, token, ',');
-	entityData.supplyCost = stoi(token);
+	entityData.supplyCost = stod(token);
 	getline(iss, token, ',');
 	entityData.supplyProvided = stoi(token);
 	getline(iss, token, ',');
